DelayedResult should have been moved with PipelineResults in the first place.
